ST-369: Refactor view model and move messages back to the vm.
